src/seminars/first/hw/Shop.java
package seminars.first.hw;

import java.util.ArrayList;
import java.util.Comparator;
import java.util.List;

public class Shop {
    private List<Product> products;

    // Геттеры, сеттеры:
    public List<Product> getProducts() {
        return products;
    }

    public void setProducts(List<Product> products) {
        this.products = products;
    }

    public Shop(List<Product> products) {
        this.products = products;
    }

    /**
     * @return отсортированный по возрастанию и цене список продуктов
     */
    public List<Product> getSortedListProducts() {
        List<Product> clone_list = new ArrayList<>(products);
        clone_list.sort(Comparator.comparing(Product::getCost));
        return clone_list;
    }

    /**
     * @return самый дорогой продукт
     */
    public Product getMostExpensiveProduct() {
        return products.stream().max(Comparator.comparing(Product::getCost)).get();
    }
}